+++ testcases/while.c
@@ -1,7 +1,7 @@
 int main()
 {
-    int i = 0;
-    int ret = 0;
+    long long i = 0;
+    long long ret = 0;
     while (i < 10) {
         ret += i;
         i++;
